requirements.txt
contourpy==1.1.0
cycler==0.11.0
fonttools==4.42.1
joblib==1.3.2
keras==2.13.1
kiwisolver==1.4.5
matplotlib==3.7.2
mysql-connector-python==8.0.33
numpy==1.25.2
opencv-python==4.8.0.76
packaging==23.1
pandas==2.1.0
Pillow==10.0.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2023.3.post1
scikit-learn==1.3.0
scipy==1.11.2
six==1.16.0
threadpoolctl==3.2.0
tzdata==2023.3
